requirements.txt
kafka-python
yfinance
chromadb
sentence-transformers
streamlit
groq
python-dotenv
pandas
plotly          
vaderSentiment   
watchdog